+++ category/UIKit/UIStoryboard+RFKit.h
@@ -3,7 +3,7 @@
     RFKit
 
     Copyright (c) 2013 BB9z
-    http://github.com/bb9z/RFKit
+    https://github.com/bb9z/RFKit
 
     The MIT License (MIT)
     http://www.opensource.org/licenses/mit-license.php