+++ dout.h
@@ -5,7 +5,7 @@
 	ver 1.3.1
  
     Copyright (c) 2012-2013 BB9z
-    http://github.com/bb9z/RFKit
+    https://github.com/bb9z/RFKit
 
     The MIT License (MIT)
     http://www.opensource.org/licenses/mit-license.php